Fix: Check if keymap-desktop.json exists

+++ src/index.ts
@@ -175,14 +175,19 @@ joplin.plugins.register({
 
         // Backup Keymap
         const profileDir = await joplin.settings.globalValue("profileDir");
-        try {
-          fs.copyFileSync(
-            profileDir + "/keymap-desktop.json",
-            backupPath + "/keymap-desktop.json"
-          );
-        } catch (e) {
-          showError("Backup error", e);
-          throw e;
+        if (fs.existsSync(profileDir + "/keymap-desktop.json")) {
+          try {
+            fs.copyFileSync(
+              profileDir + "/keymap-desktop.json",
+              backupPath + "/keymap-desktop.json"
+            );
+          } catch (e) {
+            showError("Backup error", e);
+            throw e;
+          }
+        }
+        else{
+          console.info("No keymap file '" + profileDir + "/keymap-desktop.json");
         }
 
         await joplin.settings.setValue("lastBackup", backupDate.getTime());
